rm devToken helper and move back into devSetup file
