+++ deploy/01_devSetup.ts
@@ -1,6 +1,5 @@
 import type { HardhatRuntimeEnvironment } from 'hardhat/types';
 import { ethers } from 'hardhat';
-import deployDevToken from './helpers/devToken';
 
 const { parseUnits } = ethers.utils;
 
@@ -12,6 +11,45 @@ const func = async function (env: HardhatRuntimeEnvironment) {
 
   await deployDevToken(env, 'LUSD', 'MockLUSD');
 };
+
+async function deployDevToken(
+  env: HardhatRuntimeEnvironment,
+  name: string,
+  contract: string,
+) {
+  const { deployer, alice, bob, carol } = await env.getNamedAccounts();
+  const { deploy, execute, getOrNull, read } = env.deployments;
+
+  const isDeployed = await getOrNull(name);
+
+  if (!isDeployed) {
+    const deployment = await deploy(name, {
+      contract,
+      from: deployer,
+      args: [0],
+      log: true,
+    });
+
+    if (process.env.NODE_ENV !== 'test')
+      await env.tenderly.persistArtifacts({
+        name,
+        address: deployment.address,
+      });
+
+    if (process.env.NODE_ENV !== 'test') {
+      for (let account of [deployer, alice, bob, carol]) {
+        const decimals = await read(name, 'decimals');
+        await execute(
+          name,
+          { from: account },
+          'mint',
+          account,
+          parseUnits('100000', decimals),
+        );
+      }
+    }
+  }
+}
 
 // Deploy only to hardhat and ropsten
 func.skip = async (hre: HardhatRuntimeEnvironment) =>
